face-recognition/requirements.txt
boto3
ffmpeg
ffmpeg-python
opencv-python
imutils
facenet-pytorch
Pillow
torch